Adding documentation and creating utils file for shared functions
